+++ codice/settings.py
@@ -52,6 +52,8 @@ LOCAL_APPS = [
     'authentication',
     'dashboard',
     'repos',
+    'commits',
+    'developers',
 ]
 
 INSTALLED_APPS = DJANGO_APPS + THIRD_PARTY_APPS + LOCAL_APPS