+++ CHANGELOG.md
@@ -15,3 +15,5 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - [codice-8] remove repos
 - [codice-7] add gravatar support
 - [codice-10] clone repo
+- [codice-3] analyze repo
+